Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,19 +4,29 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.resources.exceptions.DatabaseException;
+import com.educandoweb.course.resources.exceptions.ResourceExceptionHandler;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 
 @Service
 public class UserService {
+
+    private final ResourceExceptionHandler resourceExceptionHandler;
 	
 	
 	@Autowired
 	private UserRepository repository;
+
+    UserService(ResourceExceptionHandler resourceExceptionHandler) {
+        this.resourceExceptionHandler = resourceExceptionHandler;
+    }
 	
 	public List<User> findAll(){
 		return repository.findAll();
@@ -33,7 +43,20 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (repository.existsById(id)) {
+				repository.deleteById(id);
+			}
+			else {
+				throw new ResourceNotFoundException(id);
+			}
+			
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
